+++ src/test/resources/com/atomist/rug/test/gherkin/handler/event/PassingFeature2Steps1.ts
@@ -3,7 +3,10 @@ import * as node from "../../../handlers/event/Nodes"
 
 When("i call you", world => {
    world.registerHandler("ReturnsAMessage")
+   world.registerHandler("AngryHandler")
    world.sendEvent(new node.Commit)
+   world.sendEvent(new node.Commit)
+   world.sendEvent(new node.GitHubId("myId"))
 })
 Then("you call me", world => {
     return world.message() != null