+++ src/test/scala/com/atomist/rug/runtime/js/JavaScriptEventHandlerTest.scala
@@ -46,9 +46,6 @@ object JavaScriptEventHandlerTest {
        |    })
        |    plan.add(message);
        |
-       |    const jsonMessage = new Message({ value: "message2"})
-       |    plan.add(jsonMessage);
-       |
        |    plan.add({ instruction: {
        |                 kind: "execute",
        |                 name: "HTTP",
@@ -125,8 +122,9 @@ object JavaScriptEventHandlerTest {
        |@EventHandler("BuildHandler", "Handles a Build event", new PathExpression<TreeNode,TreeNode>("/issue"))
        |@Tags("github", "build")
        |class SimpleHandler implements HandleEvent<TreeNode,TreeNode> {
-       |  handle(event: Match<TreeNode, TreeNode>): Message{
-       |     return new Message("woot").withCorrelationId("dude").withNode(event.root());
+       |
+       |  handle(event: Match<TreeNode, TreeNode>) {
+       |     return new Plan().add(new Message("woot").withCorrelationId("dude").withNode(event.root()));
        |  }
        |}
        |export let handler = new SimpleHandler();
@@ -141,8 +139,8 @@ object JavaScriptEventHandlerTest {
        |@EventHandler("BuildHandler", "Handles a Build event", new PathExpression<TreeNode,TreeNode>("/nomatch"))
        |@Tags("github", "build")
        |class SimpleHandler implements HandleEvent<TreeNode,TreeNode> {
-       |  handle(event: Match<TreeNode, TreeNode>): Message{
-       |     return new Message("woot").withCorrelationId("dude").withNode(event.root());
+       |  handle(event: Match<TreeNode, TreeNode>){
+       |     return new Plan().add(new Message("woot").withCorrelationId("dude").withNode(event.root()));
        |  }
        |}
        |export let handler = new SimpleHandler();
@@ -155,7 +153,9 @@ class JavaScriptEventHandlerTest extends FlatSpec with Matchers with DiagrammedA
   import JavaScriptEventHandlerTest._
 
   it should "extract and run an event handler" in {
-    val rugArchive = TypeScriptBuilder.compileWithModel(SimpleFileBasedArtifactSource(JavaScriptEventHandlerTest.reOpenCloseIssueProgram))
+    val rugArchive = TypeScriptBuilder.compileWithModel(
+      SimpleFileBasedArtifactSource(
+        JavaScriptEventHandlerTest.reOpenCloseIssueProgram))
     val finder = new JavaScriptEventHandlerFinder()
     val handlers = finder.find(new JavaScriptContext(rugArchive))
     handlers.size should be(1)
@@ -176,8 +176,7 @@ class JavaScriptEventHandlerTest extends FlatSpec with Matchers with DiagrammedA
               Nil, None
             )), None)
         ),
-        Some("w00t")),
-      Message(JsonBody("[value=message2]"), Nil, None)
+        Some("w00t"))
     )
     val actualPlan = handler.handle(LocalRugContext(TestTreeMaterializer), SysEvent)
 
@@ -259,7 +258,8 @@ class JavaScriptEventHandlerTest extends FlatSpec with Matchers with DiagrammedA
   }
 
   it should "allow a correlation id and treenode to be added to a message" in {
-    val rugArchive = TypeScriptBuilder.compileWithModel(SimpleFileBasedArtifactSource(JavaScriptEventHandlerTest.eventHandlerWithTreeNode))
+    val rugArchive = TypeScriptBuilder.compileWithModel(
+      SimpleFileBasedArtifactSource(JavaScriptEventHandlerTest.eventHandlerWithTreeNode))
     val finder = new JavaScriptEventHandlerFinder()
     val handlers = finder.find(new JavaScriptContext(rugArchive))
     handlers.size should be(1)